monitor/user_management/requirements.txt
fastapi
uvicorn
tortoise-orm
redis
cryptography
pydantic_settings
pika
aiomysql

rich